update urls and links
