+++ grad/urls.py
@@ -6,16 +6,17 @@ ST_ID = '(?P<st_id>' + ID_RE + ')'
 
 grad_patterns = [ # prefix /grad/
     url(r'^$', grad_views.index, name='index'),
+    url(r'^active$', grad_views.active_students, name='active_students'),
     url(r'^browse$', grad_views.browse, name='browse'),
-    #url(r'^import$', grad_views.import_applic, name='import_applic'),
     url(r'^config$', grad_views.config, name='config'),
     url(r'^reports$', grad_views.reports, name='reports'),
+    #url(r'^import$', grad_views.import_applic, name='import_applic'),
     url(r'^progress_reports', grad_views.progress_reports, name='progress_reports'),
     url(r'^search$', grad_views.search, name='search'),
-    url(r'^search_index$', grad_views.search_index, name='search_index'),
     url(r'^search/save$', grad_views.save_search, name='save_search'),
     url(r'^search/delete$', grad_views.delete_savedsearch, name='delete_savedsearch'),
     url(r'^qs', grad_views.quick_search, name='quick_search'),
+    url(r'^add_supervisors', grad_views.add_supervisors, name='add_supervisors'),
 
     url(r'^program/new$', grad_views.new_program, name='new_program'),
     url(r'^program/(?P<program_id>\d+)/edit', grad_views.edit_program, name='edit_program'),
@@ -37,7 +38,6 @@ grad_patterns = [ # prefix /grad/
     url(r'^funding/(?P<semester_name>\d{4})/download_scholarships', grad_views.funding_report_download, kwargs={'type': 'scholarships'}, name='funding_report_scholarships'),
     url(r'^funding/(?P<semester_name>\d{4})/download_other', grad_views.funding_report_download, kwargs={'type': 'other'}, name='funding_report_other'),
     url(r'^financials_report', grad_views.financials_report, name='financials_report'),
-    url(r'^add_supervisors', grad_views.add_supervisors, name='add_supervisors'),
     
     url(r'^' + GRAD_SLUG + '/$', grad_views.view, name='view'),
     url(r'^' + GRAD_SLUG + '/update_note$', grad_views.update_note, name='update_note'),
