+++ media/style/grad.css
@@ -230,3 +230,13 @@ abbr .title {
 .submit_button {
   padding-top: 50px;
 }
+
+.supervisor_notification {
+  color: #777;
+  font-size: 1em;
+  font-family: sans-serif;
+  background-color: #eee;
+  padding-left: 15px;
+  border: 1px solid #bbb;
+  border-radius: 8px;
+}
