+++ templates/dashboard/index.html
@@ -30,7 +30,7 @@ $(document).ready(function() {
     {% if 'SPAC' in roles %}<li><a href="{% url "space:index" %}"><i class="fa fa-li fa-building"></i> Space Management</a></li>{% endif %}
     {% if 'RELA' in roles %}<li><a href="{% url "relationships:index" %}"><i class="fa fa-li fa-address-book"></i> Relationships </a></li>{% endif %}
 
-    {% if 'GRAD' in roles or 'GRPD' in roles %}<li class="newsec"><a href="{% url "grad:search_index" %}"><i class="fa fa-li fa-group"></i> Grad Students</a></li>{% endif %}
+    {% if 'GRAD' in roles or 'GRPD' in roles %}<li class="newsec"><a href="{% url "grad:index" %}"><i class="fa fa-li fa-group"></i> Grad Students</a></li>{% endif %}
 
     {% if 'FUND' in roles %}<li class="newsec"><a href="{% url "ra:dashboard" %}"><i class="fa fa-li fa-money"></i> Research Associates</a></li>{% endif %}
     {% if 'ADMN' in roles or 'TAAD' in roles or is_instructor %}<li><a href="{% url "tugs:all_tugs_admin" %}"><i class="fa fa-li fa-clock-o"></i> Time Use Guidelines</a></li>{% endif %}
