+++ templates/grad/add_supervisors.html
@@ -25,12 +25,12 @@ $(document).ready(function() {
 </script>
 {% endblock %}
 
-{% block title %}Add Committee Members{% endblock %}
-{% block h1 %}Add Committee Members{% endblock %}
+{% block title %}Bulk Add Committee Members{% endblock %}
+{% block h1 %}Bulk Add Committee Members{% endblock %}
 
 {% block subbreadcrumbs %}
-<li><a href="{% url "grad:search_index" %}">Grad</a></li>
-<li>Add Committee Members</li>
+<li><a href="{% url "grad:index" %}">Grad</a></li>
+<li>Bulk Add Committee Members</li>
 {% endblock %}
 
 {% block content %}
