+++ templates/grad/config.html
@@ -18,9 +18,9 @@
 <div class="actionlist">
     <div class="btn pressed">Configuration</div>
     <a class="btn" href="{% url "grad:reports" %}">Reports</a>
-    <a class="btn" href="{% url "grad:search_index" %}">Search</a>
+    <a class="btn" href="{% url "grad:index" %}">Search</a>
     <a class="btn" href="{% url "grad:browse" %}">Browse All Students</a>
-    <a class="btn" href="{% url "grad:index" %}">Active Students</a>
+    <a class="btn" href="{% url "grad:active_students" %}">Active Students</a>
 </div>
 <table id="masters" class="display">
     <thead>
