+++ templates/grad/reports.html
@@ -18,9 +18,9 @@
 <div class="actionlist">
     <a class="btn" href="{% url "grad:config" %}">Configuration</a>
     <div class="btn pressed">Reports</div>
-    <a class="btn" href="{% url "grad:search_index" %}">Search</a>
+    <a class="btn" href="{% url "grad:index" %}">Search</a>
     <a class="btn" href="{% url "grad:browse" %}">Browse All Students</a>
-    <a class="btn" href="{% url "grad:index" %}">Active Students</a>
+    <a class="btn" href="{% url "grad:active_students" %}">Active Students</a>
 </div>
 <table id="masters" class="display">
     <thead>
